Add option to keep Toastify running without Spotify (#49)
Added settings:
- `CanRunWithoutSpotify`
- `StartSpotifyWithToastify`
The latter only appies if the former is enabled.

The settings are still not modifiable from the UI.
This feature requires Admin privileges; this requirement should be made clear in the UI. If Toastify is started withoud admin privileges, the settings are ignored and the default behaviour will apply.